Refactor tests: remove unused classes, update PersistenceShowcase

+++ BYTPRO.Test/Showcase/PersistenceShowcase.cs
@@ -1,7 +1,4 @@
-﻿using BYTPRO.Data.JsonUoW;
-using BYTPRO.Data.Models;
-using BYTPRO.Data.Models.People;
-using BYTPRO.Data.Models.Sales.Orders;
+﻿using BYTPRO.Data.Models.People;
 using BYTPRO.JsonEntityFramework.Context;
 using BYTPRO.JsonEntityFramework.Extensions;
 using Xunit.Abstractions;
@@ -11,82 +8,39 @@ namespace BYTPRO.Test.Showcase;
 public class PersistenceShowcase(ITestOutputHelper testOutputHelper)
 {
     private static string DbRoot => $"{Directory.GetCurrentDirectory()}/Db";
-    
+
     static PersistenceShowcase()
     {
-        if (!Directory.Exists(DbRoot)) Directory.CreateDirectory(DbRoot);
-        
+        if (Directory.Exists(DbRoot)) Directory.Delete(DbRoot, true);
+        Directory.CreateDirectory(DbRoot);
+
         var context = new JsonContextBuilder()
             .AddJsonEntity<Customer>()
-                .WithFileName("customer")
-                .BuildEntity()
-            .AddJsonEntity<Order>()
-                .WithFileName("order")
-                .BuildEntity()
+            .WithFileName("customers")
+            .BuildEntity()
             .WithRoot(new DirectoryInfo(DbRoot))
             .Build();
 
         JsonContext.SetContext(context);
     }
-    
-    [Fact]
-    public async Task CreatePerson()
-    {
-        var context = JsonContext.Context;
-        
-        var person = new Customer(
-            1,
-            "Artiom", 
-            "Bezkorovainyi",
-            "+48000000000", 
-            "s30000@pjwstk.edu.pl", 
-            "12345678",
-            DateTime.Now
-        );
 
-        await new JsonUnitOfWork(context).SaveChangesAsync();
-        
-        testOutputHelper.WriteLine($"After create {Person.All.ToJson()}");
-    }
-        
-    
+
     [Fact]
     public async Task TestPersistence()
     {
-        var context = JsonContext.Context;
-        var uow = new JsonUnitOfWork(context);
-        
-        var person = new Customer(
+        var customer = new Customer(
             1,
-            "Artiom", 
+            "Artiom",
             "Bezkorovainyi",
-            "+48000000000", 
-            "s30000@pjwstk.edu.pl", 
+            "+48000000000",
+            "s30000@pjwstk.edu.pl",
             "12345678",
             DateTime.Now
         );
 
-        await uow.SaveChangesAsync();
-        
-        testOutputHelper.WriteLine($"After add {Person.All.ToJson()}");
-        
-        Customer.Remove(person);
-        
-        await uow.SaveChangesAsync();
-        
-        testOutputHelper.WriteLine($"After delete {Person.All.ToJson()}");
-    }
-    
-    [Fact]
-    public void ShowPersons()
-    {
-        var context = JsonContext.Context;
-        testOutputHelper.WriteLine(context.GetTable<Customer>().ToJson());
-    }
-    
-    [Fact]
-    public void RemoveDb()
-    {
-        Directory.Delete(DbRoot, true);
+        await JsonContext.Context.SaveChangesAsync();
+
+        testOutputHelper.WriteLine($"Persons: {Person.All.ToJson()}");
+        testOutputHelper.WriteLine($"\nCustomers: {Customer.All.ToJson()}");
     }
 }
